Update wild-theme-shell and fix config name
We need to avoid root's go.mod and go.sum being modified except for
the actual dependency on wild-theme-shell. We therefore rename
tests/config/hugo.toml as tests/config/test-config.toml, which Hugo does
not automatically use, and update our scripts to point to this config
for testing builds and the demo site.

Signed-off-by: Daniel F. Dickinson <dfdpublic@wildtechgarden.ca>